+++ ui/components/activity-timeline/flavors/call/index.react.example.jsx
@@ -22,12 +22,12 @@ export let states = [
     element:
       <div className="demo-only">
         <span className="slds-assistive-text">Call</span>
-        <MediaObject figureRight={ <TimelineActions /> }>
+        <MediaObject figureRight={ <TimelineActions title="More Options for Call, Mobile conversation" assistiveText="More Options for Call, Mobile conversation" /> }>
           <MediaObject className="slds-media--timeline slds-timeline__media--call" figureLeft={ <TimelineIcon symbol="log_a_call" /> } figureLeftClassName="slds-timeline__icon">
             <h3 className="slds-truncate" title="Mobile conversation on Monday">
               <a href="javascript:void(0);">Mobile conversation on Monday</a>
             </h3>
-            <p className="slds-truncate">Lei seemed interested in closing this deal quickly! Let&rsquo;s move.</p>
+            <p className="slds-truncate" title="Lei seemed interested in closing this deal quickly! Let&rsquo;s move.">Lei seemed interested in closing this deal quickly! Let&rsquo;s move.</p>
             <ul className="slds-list--horizontal slds-wrap">
               <li className="slds-m-right--large">
                 <span className="slds-text-title">Name:</span>