+++ ui/components/button-icons/flavors/base/index.react.example.jsx
@@ -24,7 +24,7 @@ let Demo = props =>
 
 export let ButtonIcon = props =>
   <button
-    className={className('slds-button', props.className, props.stateful ? props.selected ? 'slds-is-selected' : 'slds-not-selected' : null)}
+    className={className('slds-button', props.className, props.selected ? 'slds-is-selected' : null)}
     disabled={ props.disabled }
     role={ props.role }
     aria-describedby={ props['aria-describedby'] }
@@ -34,8 +34,8 @@ export let ButtonIcon = props =>
     aria-hidden={ props['aria-hidden'] }
     aria-pressed={ props['aria-pressed'] }
     aria-live={ props.assertive ? 'assertive' : props['aria-live'] }
-    title={ props.title }
-    tabIndex={ props.tabIndex } >
+    tabIndex={ props.tabIndex }
+    title={ props.title || 'settings' } >
     <SvgIcon
       className={className('slds-button__icon', props.iconClassName)}
       sprite="utility"
@@ -57,7 +57,7 @@ export let states = [
   {
     id: 'button-icon-error',
     label: 'Error',
-    element: <Demo><ButtonIcon className="slds-button--icon-error" symbol="warning" /></Demo>
+    element: <Demo><ButtonIcon className="slds-button--icon-error" symbol="warning" title="Warning" assistiveText="Warning" /></Demo>
   },
   {
     id: 'button-icon-disabled',