+++ ui/components/button-icons/index.docs.jsx
@@ -24,7 +24,7 @@ export default (
   <ComponentDocs>
     <h4 className="site-text-heading--label">Accessibility</h4>
     <p>
-      If an icon button doesn’t include a label, use a span with <CodeClass className="assistive-text" /> to describe the icon for screen readers.
+      If an icon button doesn’t include a label, use a <code>title</code> attribute to show on hover for sighted users, and a span with <CodeClass className="assistive-text" /> to describe the icon for screen readers.
     </p>
   </ComponentDocs>
 );