+++ ui/components/docked-composer/flavors/email/index.react.example.jsx
@@ -65,41 +65,50 @@ export default (
               <ButtonIcon
                 className="slds-button--icon-border-filled"
                 symbol="bold"
-                assistiveText="Bold" />
+                assistiveText="Bold"
+                title="Bold" />
               <ButtonIcon
                 className="slds-button--icon-border-filled"
                 symbol="italic"
-                assistiveText="Italic" />
+                assistiveText="Italic"
+                title="Italic" />
               <ButtonIcon
                 className="slds-button--icon-border-filled"
                 symbol="underline"
-                assistiveText="Underline" />
+                assistiveText="Underline"
+                title="Underline" />
             </div>
             <div className="slds-button-group" role="group">
               <ButtonIcon
                 className="slds-button--icon-border-filled"
                 symbol="richtextnumberedlist"
-                assistiveText="Numbered List" />
+                assistiveText="Numbered List"
+                title="Numbered List" />
               <ButtonIcon
                 className="slds-button--icon-border-filled"
                 symbol="richtextbulletedlist"
-                assistiveText="Bulleted List" />
+                assistiveText="Bulleted List"
+                title="Bulleted List" />
               <ButtonIcon
                 className="slds-button--icon-border-filled"
                 symbol="left_align_text"
-                assistiveText="Left Align Text" />
+                assistiveText="Left Align Text"
+                title="Left Align Text" />
               <ButtonIcon
                 className="slds-button--icon-border-filled"
                 symbol="center_align_text"
-                assistiveText="Center Align Text" />
+                assistiveText="Center Align Text"
+                title="Center Align Text" />
               <ButtonIcon
                 className="slds-button--icon-border-filled"
                 symbol="right_align_text"
-                assistiveText="Right Align Text" />
+                assistiveText="Right Align Text"
+                title="Right Align Text" />
               <ButtonIcon
                 className="slds-button--icon-border-filled"
                 symbol="link"
-                assistiveText="Link" />
+                assistiveText="Link"
+                title="Link" />
             </div>
           </div>
         </div>
@@ -111,23 +120,28 @@ export default (
           <ButtonIcon
             className="slds-button--icon-container"
             symbol="link"
-            assistiveText="Attach File" />
+            assistiveText="Attach File"
+            title="Attach File" />
           <ButtonIcon
             className="slds-button--icon-container"
             symbol="insert_template"
-            assistiveText="Insert Template" />
+            assistiveText="Insert Template"
+            title="Insert Template" />
           <ButtonIcon
             className="slds-button--icon-container"
             symbol="insert_tag_field"
-            assistiveText="Insert HTML" />
+            assistiveText="Insert HTML"
+            title="Insert HTML" />
           <ButtonIcon
             className="slds-button--icon-container"
             symbol="preview"
-            assistiveText="Preview" />
+            assistiveText="Preview"
+            title="Preview" />
           <ButtonIcon
             className="slds-button--icon-container"
             symbol="delete"
-            assistiveText="Delete" />
+            assistiveText="Delete"
+            title="Delete" />
           <button className="slds-button slds-button--brand">Send</button>
         </div>
       </DockedComposerPanelFooter>