+++ ui/components/trees/flavors/grid/index.react.example.jsx
@@ -21,7 +21,7 @@ import Checkbox from 'ui/components/forms/flavors/checkbox/index.react';
 let Expanded = props =>
 <tr className="slds-hint-parent" id="tree0-node1-0" role="row" aria-level="2">
   <th data-label="Account Name" scope="row" className="slds-tree__item">
-    <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" disabled />
+    <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" disabled />
     <div className="slds-truncate" title="Acme Corporation (Oakland)"><a href="javascript:void(0);">Acme Corporation (Oakland)</a></div>
   </th>
   <td data-label="Employees"><div className="slds-truncate" title="745">745</div></td>
@@ -41,7 +41,7 @@ let Default = props =>
 <tbody>
   <tr className="slds-hint-parent" id="tree0-node0" role="row" aria-level="1">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" disabled />
+      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" disabled />
       <div className="slds-truncate" title="Rewis Inc"><a href="javascript:void(0);">Rewis Inc</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="3,100">3,100</div></td>
@@ -58,7 +58,7 @@ let Default = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node1" role="row" aria-level="1" aria-owns={props.isExpanded ? 'tree0-node1-0' : null} aria-expanded={props.isExpanded ? true : false}>
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon aria-controls="tree0-node1" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" />
+      <ButtonIcon aria-controls="tree0-node1" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" />
       <div className="slds-truncate" title="Acme Corporation"><a href="javascript:void(0);">Acme Corporation</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="10,000">10,000</div></td>
@@ -76,7 +76,7 @@ let Default = props =>
   {props.additionalItem}
   <tr className="slds-hint-parent" id="tree0-node2" role="row" aria-level="1" aria-expanded="false">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon aria-controls="tree0-node2" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" />
+      <ButtonIcon aria-controls="tree0-node2" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" />
       <div className="slds-truncate" title="Rohde Enterprises"><a href="javascript:void(0);">Rohde Enterprises</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="6,000">6,000</div></td>
@@ -93,7 +93,7 @@ let Default = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node3" role="row" aria-level="1">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" disabled />
+      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" disabled />
       <div className="slds-truncate" title="Cheese Corp"><a href="javascript:void(0);">Cheese Corp</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="1,234">1,234</div></td>
@@ -115,7 +115,7 @@ let DeepNesting = props =>
 <tbody>
   <tr className="slds-hint-parent" id="tree0-node0" role="row" aria-level="1">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" disabled />
+      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" disabled />
       <div className="slds-truncate" title="Rewis Inc"><a href="javascript:void(0);">Rewis Inc</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="3,100">3,100</div></td>
@@ -132,7 +132,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node1" role="row" aria-level="1" aria-owns="tree0-node1-0 tree0-node1-1" aria-expanded="true">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon aria-controls="tree0-node1" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" />
+      <ButtonIcon aria-controls="tree0-node1" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" />
       <div className="slds-truncate" title="Acme Corporation"><a href="javascript:void(0);">Acme Corporation</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="10,000">10,000</div></td>
@@ -149,7 +149,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node1-0" role="row" aria-level="2" aria-owns="tree0-node1-0-0 tree0-node1-0-1" aria-expanded="true">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon aria-controls="tree0-node1-0" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" />
+      <ButtonIcon aria-controls="tree0-node1-0" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" />
       <div className="slds-truncate" title="Acme Corporation (Bay Area)"><a href="javascript:void(0);">Acme Corporation (Bay Area)</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="3,000">3,000</div></td>
@@ -166,7 +166,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node1-0-0" role="row" aria-level="3">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" disabled />
+      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" disabled />
       <div className="slds-truncate" title="Acme Corporation (Oakland)"><a href="javascript:void(0);">Acme Corporation (Oakland)</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="745">745</div></td>
@@ -183,8 +183,8 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node1-0-1" role="row" aria-level="3">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" disabled />
-      {/*<ButtonIcon className="slds-m-right--x-small slds-is-disabled" flavor="icon-bare" iconFlavor="small" sprite="utility" symbol="chevronright" assistiveText="Toggle" disabled />*/}
+      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" disabled />
+      {/*<ButtonIcon className="slds-m-right--x-small slds-is-disabled" flavor="icon-bare" iconFlavor="small" sprite="utility" symbol="chevronright" assistiveText="Toggle" title="Toggle" disabled />*/}
       <div className="slds-truncate" title="Acme Corporation (San Francisco)"><a href="javascript:void(0);">Acme Corporation (San Francisco)</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="578">578</div></td>
@@ -201,7 +201,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node1-1" role="row" aria-level="2" aria-owns="tree0-node1-1-0 tree0-node1-1-1" aria-expanded="true">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon aria-controls="tree0-node1-1" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" />
+      <ButtonIcon aria-controls="tree0-node1-1" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" />
       <div className="slds-truncate" title="Acme Corporation (East)"><a href="javascript:void(0);">Acme Corporation (East)</a></div>
   </th>
     <td data-label="Employees"><div className="slds-truncate" title="430">430</div></td>
@@ -218,7 +218,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node1-1-0" role="row" aria-level="3">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" disabled />
+      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" disabled />
       <div className="slds-truncate" title="Acme Corporation (NY)"><a href="javascript:void(0);">Acme Corporation (NY)</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="1,210">1,210</div></td>
@@ -235,7 +235,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node1-1-1" role="row" aria-level="3" aria-owns="tree0-node1-0-1-0" aria-expanded="true">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon aria-controls="tree0-node1-1-1" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" />
+      <ButtonIcon aria-controls="tree0-node1-1-1" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" />
       <div className="slds-truncate" title="Acme Corporation (VA)"><a href="javascript:void(0);">Acme Corporation (VA)</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="410">410</div></td>
@@ -252,7 +252,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node1-0-1-0" role="row" aria-level="4" aria-owns="tree0-node1-0-1-0-0" aria-expanded="true">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon aria-controls="tree0-node1-0-1-0" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" />
+      <ButtonIcon aria-controls="tree0-node1-0-1-0" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" />
       <div className="slds-truncate" title="Allied Technologies"><a href="javascript:void(0);">Allied Technologies</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="390">390</div></td>
@@ -269,7 +269,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node1-0-1-0-0" role="row" aria-level="5">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" disabled />
+      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" disabled />
       <div className="slds-truncate" title="Allied Technologies (UV)"><a href="javascript:void(0);">Allied Technologies (UV)</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="270">270</div></td>
@@ -286,7 +286,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node2" role="row" aria-level="1" aria-owns="tree0-node2-0" aria-expanded="true">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon aria-controls="tree0-node2" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" />
+      <ButtonIcon aria-controls="tree0-node2" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" />
       <div className="slds-truncate" title="Rohde Enterprises"><a href="javascript:void(0);">Rohde Enterprises</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="6,000">6,000</div></td>
@@ -303,7 +303,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node2-0" role="row" aria-level="2">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" disabled />
+      <ButtonIcon className="slds-button--icon slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" disabled />
       <div className="slds-truncate" title="Rohde Enterprises (UCA)"><a href="javascript:void(0);">Rohde Enterprises (UCA)</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="2,540">2,540</div></td>
@@ -320,7 +320,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node3" role="row" aria-level="1" aria-owns="tree0-node3-0" aria-expanded="true">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon aria-controls="tree0-node3" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" />
+      <ButtonIcon aria-controls="tree0-node3" className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small" iconClassName="slds-button__icon--small" symbol="chevronright" assistiveText="Toggle" title="Toggle" />
       <a href="javascript:void(0);">Tech Labs</a>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="1,856">1,856</div></td>
@@ -337,7 +337,7 @@ let DeepNesting = props =>
   </tr>
   <tr className="slds-hint-parent" id="tree0-node3-0" role="row" aria-level="2">
     <th data-label="Account Name" scope="row" className="slds-tree__item">
-      <ButtonIcon className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevrondown" assistiveText="Toggle" disabled />
+      <ButtonIcon className="slds-button--icon slds-button--icon-x-small slds-m-right--x-small slds-is-disabled" iconClassName="slds-button__icon--small" symbol="chevrondown" assistiveText="Toggle" title="Toggle" disabled />
       <div className="slds-truncate" title="Opportunity Resources Inc"><a href="javascript:void(0);">Opportunity Resources Inc</a></div>
     </th>
     <td data-label="Employees"><div className="slds-truncate" title="1,934">1,934</div></td>
@@ -364,7 +364,7 @@ let TreeGrid = props =>
     <tr className="slds-text-title--caps">
       <th className="slds-cell-buffer--left" scope="col">
         <div className="slds-grid slds-grid--vertical-align-center">
-          <ButtonIcon className="slds-button--icon-border-filled slds-button--icon-x-small slds-m-right--x-small slds-shrink-none slds-table--tree__toggle" iconClassName="slds-button__icon--small" symbol="chevrondown" assistiveText="Toggle" />
+          <ButtonIcon className="slds-button--icon-border-filled slds-button--icon-x-small slds-m-right--x-small slds-shrink-none slds-table--tree__toggle" iconClassName="slds-button__icon--small" symbol="chevrondown" assistiveText="Toggle" title="Toggle" />
           <div className="slds-truncate" title="Account Name">Account Name</div>
         </div>
       </th>