+++ RELEASENOTES.md
@@ -5,6 +5,17 @@
 
 <!-- ## [Unreleased] -->
 
+## Release [2.1.4] - November 7, 2016
+
+**FIXED**
+
+The following issues have now been resolved:
+
+- Title attributes were added to all button icons and truncated text for accessibility
+- Added line-height fix for `.slds-button—icon-border-inverse`
+- Amended `.slds-button--icon` documentation
+- Removed unnecessary `.slds-not-selected` class from stateful `.slds-button--icon` markup and adjusted documentation
+
 ## Release [2.1.3] - September 21, 2016
 
 **FIXED**
@@ -749,7 +760,8 @@ The following features are being marked as deprecated in this release and will b
 
 Lots of bug fixes and TLC.
 
-[Unreleased]: https://github.com/salesforce-ux/design-system-internal/compare/v2.1.2...winter-17
+[Unreleased]: https://github.com/salesforce-ux/design-system-internal/compare/v2.1.4...winter-17
+[2.1.4]: https://github.com/salesforce-ux/design-system/compare/v2.1.3...v2.1.4
 [2.1.3]: https://github.com/salesforce-ux/design-system/compare/v2.1.2...v2.1.3
 [2.1.2]: https://github.com/salesforce-ux/design-system/compare/v2.0.3...v2.1.2
 [2.0.3]: https://github.com/salesforce-ux/design-system/compare/v2.0.2...v2.0.3